Add RosenbergStrongPureAlgorithm.
This is just like the existing Rosenberg-Strong-based algorithms, but doesn't add 3 to X or Y, and doesn't randomize afterward at all. An advantage here is that every cell in a 256x256 grid is guaranteed to have a unique 16-bit hash code. A disadvantage is that negative X or Y aren't handled well.