Just in case, BareMinimumAlgorithm.
This is an extremely simple algorithm, and could still be adapted for more SadRogue-specific usage. It just adds X to a bitwise-rotated Y. Changing the  return line to `return (int)(x + (y << 8 | y >> 24));` may be a good choice for the specific requirements of a hash code here, since it makes every cell in a 256x256 grid have a unique 16-bit hash code. This may not matter because Dictionary and HashSet use prime modulus tables.